Added UUID to entity detail
* UUID has been added to group, resource and user detail.
* Groups and resources now can be searched by UUID.
