+++ apps/admin-gui/src/app/admin/pages/admin-user-detail-page/admin-user-detail-page.component.html
@@ -21,6 +21,9 @@
           <mat-icon>edit</mat-icon>
         </button>
       </div>
+      <div>
+        {{'ADMIN_USER.UUID' | translate}}: {{user.uuid}}
+      </div>
       <span class="mt-1 entity-info">
         {{'ADMIN_USER.ENTITY' | translate}},
         {{'ADMIN_USER.USER_TYPE' | translate}}: {{getUserType()}}
