+++ apps/admin-gui/src/app/facilities/pages/resource-detail-page/resource-detail-page.component.html
@@ -28,6 +28,9 @@
             <mat-icon>delete_forever</mat-icon>
           </button>
         </div>
+        <div>
+          {{'RESOURCE_DETAIL.UUID' | translate}}: {{resource.uuid}}
+        </div>
         <span class="mt-1 entity-info">
           {{'RESOURCE_DETAIL.ENTITY' | translate}},
           {{'RESOURCE_DETAIL.VO' | translate}}:
