+++ apps/admin-gui/src/app/vos/pages/group-detail-page/group-detail-page.component.html
@@ -28,6 +28,9 @@
             <mat-icon>delete_forever</mat-icon>
           </button>
         </div>
+        <div>
+          {{'GROUP_DETAIL.UUID' | translate}}: {{group.uuid}}
+        </div>
         <span class="mt-1 entity-info">
           {{'GROUP_DETAIL.ENTITY' | translate}},
           {{'GROUP_DETAIL.VO_NAME' | translate}}:
