+++ libs/perun/components/src/lib/groups-tree/groups-tree.component.ts
@@ -106,7 +106,8 @@ export class GroupsTreeComponent implements OnChanges {
     if (this.expandAll) {
       this.filteredGroups = this.groups.filter( option => option.name?.toLowerCase().includes(this.filterValue.toLowerCase()) ||
                                                           option.description?.toLowerCase().includes(this.filterValue.toLowerCase()) ||
-                                                          option.id.toString().includes(this.filterValue.toLowerCase()));
+                                                          option.id.toString().includes(this.filterValue.toLowerCase()) ||
+                                                          option.uuid.toLowerCase().includes(this.filterValue.toLowerCase()));
       for (const group of this.filteredGroups) {
         if (group.parentGroupId) {
           this.filteredGroups = this.filteredGroups.concat(findParent(group.parentGroupId, this.groups));
