+++ libs/perun/openapi/src/lib/model/models.ts
@@ -24,7 +24,6 @@ export * from './banOnVo';
 export * from './brand';
 export * from './candidate';
 export * from './category';
-export * from './contactGroup';
 export * from './destination';
 export * from './destinationPropagationType';
 export * from './destinationType';
@@ -48,8 +47,6 @@ export * from './inputAddApplicationMailForGroup';
 export * from './inputAddApplicationMailForVo';
 export * from './inputAddDestinationToMultipleServices';
 export * from './inputAddDestinationsDefinedByHostsOnFacility';
-export * from './inputAddFacilityContact';
-export * from './inputAddFacilityContacts';
 export * from './inputAssignResourceTagToResource';
 export * from './inputAttributeDefinition';
 export * from './inputAttributeRights';
@@ -80,8 +77,6 @@ export * from './inputFormItemData1';
 export * from './inputGetAllResourcesByResourceTag';
 export * from './inputGetPaginatedMembers';
 export * from './inputLockPublications';
-export * from './inputRemoveFacilityContact';
-export * from './inputRemoveFacilityContacts';
 export * from './inputRemoveResourceTagFromResource';
 export * from './inputSendMessage';
 export * from './inputSetBan';
