+++ libs/perun/openapi/src/lib/model/resource.ts
@@ -17,5 +17,6 @@ export interface Resource extends Auditable {
     description?: string;
     voId?: number;
     facilityId?: number;
+    uuid?: string;
 }
